wouldyouci_database/requirements.txt
joblib==0.15.1
numpy==1.18.4
pandas==1.0.4
PyMySQL==0.9.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
scikit-learn==0.23.1
scikit-surprise==1.1.0
scipy==1.4.1
six==1.15.0
sklearn==0.0
surprise==0.1
threadpoolctl==2.1.0
